confluence-engine/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# Async HTTP
httpx>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0

# Scheduling
apscheduler>=3.10.0

# Config
pyyaml>=6.0

# Broker (Phase 4)
alpaca-py>=0.21.0

# Dev/Testing
pytest>=8.0
pytest-asyncio>=0.23.0
httpx  # Also used for testing FastAPI
